+++ flag/service/service.go
@@ -2,9 +2,14 @@ package service
 
 import (
 	"github.com/giantswarm/operatorkit/v4/pkg/flag/service/kubernetes"
+
+	"github.com/giantswarm/organization-operator/flag/service/legacycredentials"
+	"github.com/giantswarm/organization-operator/flag/service/legacyorganizations"
 )
 
 // Service is an intermediate data structure for command line configuration flags.
 type Service struct {
-	Kubernetes kubernetes.Kubernetes
+	Kubernetes          kubernetes.Kubernetes
+	LegacyOrganizations legacyorganizations.LegacyOrganizations
+	LegacyCredentials   legacycredentials.LegacyCredentials
 }