+++ service/controller/key/key.go
@@ -2,6 +2,7 @@ package key
 
 import (
 	securityv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/security/v1alpha1"
+	"github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/giantswarm/microerror"
 )
 
@@ -18,4 +19,15 @@ func ToOrganization(v interface{}) (securityv1alpha1.Organization, error) {
 	c := p.DeepCopy()
 
 	return *c, nil
+}
+
+func LegacyOrganizationName(cr *securityv1alpha1.Organization) string {
+	if cr.GetAnnotations() != nil {
+		name := cr.GetAnnotations()[annotation.UIOriginalOrganizationName]
+		if len(name) > 0 {
+			return name
+		}
+	}
+
+	return cr.GetName()
 }